refactored ol map code to map.utlis.js, removed header until map is ready
