+++ src/index.html
@@ -9,10 +9,11 @@
   </head>
   <body>
     <div class="container">
-      <h1>Real-time Iss Location</h1>
+      <h1 id="header"></h1>
       <div id="map"></div>
       <div class="stars"></div>
       <div class="twinkling"></div>
+      <div class="clouds"></div>
     </div>
   </body>
 </html>
