+++ src/style.scss
@@ -25,9 +25,9 @@ html {
   align-items: center;
 }
 
-h1 {
+#header {
   position: relative;
-  z-index: 2;
+  z-index: 3;
   color: #c3cbd4;
   font-family: 'STARWARS', sans-serif;
   text-align: center;
@@ -35,28 +35,33 @@ h1 {
   line-height: 1;
   text-transform: uppercase;
   letter-spacing: 0.2em;
-  margin: 0 1em 1em;
+  margin: 0 1em 0.6em;
   text-shadow: 1px 1px 2px black;
 
   @media screen and (max-width: 768px) {
     font-size: 1.8rem;
   }
-
-  @media screen and (max-width: 400px) {
+  @media screen and (max-width: 600px) {
     font-size: 1.5rem;
+  }
+
+  @media screen and (max-width: 500px) {
+    font-size: 1.2rem;
   }
 }
 
 #map {
+  display: none;
   position: relative;
-  z-index: 2;
+  z-index: 3;
   height: 80%;
   width: 90%;
   border: 3px solid black;
 }
 
 .stars,
-.twinkling {
+.twinkling,
+.clouds {
   position: absolute;
   display: block;
   top: 0;
@@ -80,11 +85,27 @@ h1 {
   animation: move-twink-back 200s linear infinite;
 }
 
+.clouds {
+  z-index: 2;
+  background: transparent url('https://image.ibb.co/bT4N7T/clouds.png') repeat
+    top center;
+  animation: move-clouds-back 200s linear infinite;
+}
+
 @keyframes move-twink-back {
   from {
     background-position: 0 0;
   }
   to {
     background-position: -10000px 5000px;
+  }
+}
+
+@keyframes move-clouds-back {
+  from {
+    background-position: 0 0;
+  }
+  to {
+    background-position: 10000px 0;
   }
 }
